// src/models/dataSource/IDataContext.ts
import { IDataFilter, FilterConditionOperator } from '../search/IDataFilter';
import { IDataFilterConfiguration } from '../search/IDataFilterConfiguration';

/**
 * Provides useful information about the current Web Part context.
 */
export interface IDataContext {

    /**
     * The input query text if used (ex: from the page environment or search box dynamic data sources)
     */
    inputQueryText: string;

    /**
     * The current selected page number
     */
    pageNumber: number;

    /**
     * The number of items to show per page
     */
    itemsCountPerPage: number;

    /**
     * Data source paging information
     */
    paging?: {

        /**
         * The next link URL to execute. This value will reflect the value generated by the data source.
         */
        nextLinkUrl: string;

        /**
         * If the data source has dynamic paging, the available page links for the pagination component.
         * This information can be used to initialize data sources internal page links tracking when data come from cache.
         */
        pageLinks: string[];
    };

    /**
     * Information about connected filters
     */
    filters?: {

        /**
         * If connected to a 'Data Filters' Web Part, the filters configuration (will remain null otherwise)
         */
        filtersConfiguration?: IDataFilterConfiguration[];

        /**
         * The current selected filters for this data source
         */
        selectedFilters?: IDataFilter[];

        /**
         * The configured logical operator to use between filters
         */
        filterOperator?: FilterConditionOperator;

        /**
         * The connected filters Web Part instance ID
         */
        instanceId?: string;
    };
}